restructure packages, add vuu-data-react
