+++ vuu-ui/packages/vuu-data-ag-grid/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@finos/vuu-data": "0.0.26",
+    "@finos/vuu-data-react": "0.0.26",
     "@finos/vuu-filter-types": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
     "@finos/vuu-popups": "0.0.26",
