+++ vuu-ui/packages/vuu-data-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@finos/vuu-data",
+  "name": "@finos/vuu-data-react",
   "version": "0.0.26",
   "main": "src/index.ts",
   "author": "heswell",
@@ -15,7 +15,12 @@
     "@finos/vuu-protocol-types": "0.0.26"
   },
   "dependencies": {
+    "@finos/vuu-data": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
+    "@finos/vuu-popups": "0.0.26",
     "@finos/vuu-utils": "0.0.26"
+  },
+  "peerDependencies": {
+    "react": "^17.0.2"
   }
 }
