+++ vuu-ui/packages/vuu-data/src/index.ts
@@ -1,13 +1,8 @@
 export * from "./authenticate";
-export {
-  connectToServer,
-  getServerAPI,
-  makeRpcCall,
-} from "./connection-manager";
+export * from "./connection-manager";
 export type { ServerAPI } from "./connection-manager";
 export * from "./constants";
 export * from "./data-source";
-export * from "./hooks";
 export * from "./message-utils";
 export * from "./array-data-source/array-data-source";
 export * from "./json-data-source";
