+++ vuu-ui/packages/vuu-table/src/useDataSource.ts
@@ -4,11 +4,13 @@ import {
   DataSourceRow,
   DataSourceSubscribedMessage,
   SubscribeCallback,
-  VuuFeatureMessage,
-  isVuuFeatureAction,
   VuuFeatureInvocationMessage,
-  isVuuFeatureInvocation,
+  VuuFeatureMessage,
 } from "@finos/vuu-data";
+import {
+  isVuuFeatureAction,
+  isVuuFeatureInvocation,
+} from "@finos/vuu-data-react";
 import { VuuRange, VuuSortCol } from "@finos/vuu-protocol-types";
 import { getFullRange, metadataKeys, WindowRange } from "@finos/vuu-utils";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
